Move `rust-mozjs`'s build script to `mozjs-sys` (#409)
* Rename mozjs-sys jsglue static to jsapi

This also move libjsapi.a to the same place as jsapi.rs in $OUT_DIR/build

* Move mozjs's jsglue to mozjs-sys

* Cargo fmt
