+++ mozjs-sys/src/jsapi.hpp
@@ -4,8 +4,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#ifndef _JSGLUE_INCLUDED
-#define _JSGLUE_INCLUDED
+#ifndef _JSAPI_INCLUDED
+#define _JSAPI_INCLUDED
 
 #include "js/ArrayBuffer.h"
 #include "js/ArrayBufferMaybeShared.h"
