+++ mozjs-sys/src/lib.rs
@@ -13,6 +13,7 @@ extern crate libz_sys;
 mod jsimpls;
 
 // Modules with public definitions
+pub mod glue;
 pub mod jsgc;
 pub mod jsid;
 pub mod jsval;
